+++ cmd/config_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"runtime"
 
 	"testing"
 
@@ -14,7 +15,7 @@ import (
 
 func TestParseInvalidPath(t *testing.T) {
 	err := cfg.Parse("/path_do_not_exist")
-	assert.Equal(t, fmt.Errorf("Couldn't open config file : open /path_do_not_exist: no such file or directory"), err)
+	assert.Equal(t, fmt.Errorf("couldn't open config file : open /path_do_not_exist: no such file or directory"), err)
 }
 
 func TestParseInvalidYaml(t *testing.T) {
@@ -36,7 +37,7 @@ func TestParseEmptyYaml(t *testing.T) {
 	// Invalid YAML content
 	f.WriteString("---")
 	err = cfg.Parse(f.Name())
-	assert.Equal(t, fmt.Errorf("You need to configure at least one project/wildcard to poll, none given"), err)
+	assert.Equal(t, fmt.Errorf("you need to configure at least one project/wildcard to poll, none given"), err)
 }
 
 func TestParseValidConfig(t *testing.T) {
@@ -60,6 +61,7 @@ pipelines_polling_interval_seconds: 4
 on_init_fetch_refs_from_pipelines: true
 on_init_fetch_refs_from_pipelines_depth_limit: 1337
 default_refs: "^dev$"
+maximum_projects_poller_workers: 4
 
 projects:
   - name: foo/project
@@ -100,6 +102,7 @@ wildcards:
 		OnInitFetchRefsFromPipelines:           true,
 		OnInitFetchRefsFromPipelinesDepthLimit: 1337,
 		DefaultRefsRegexp:                      "^dev$",
+		MaximumProjectsPollingWorkers:          4,
 		Projects: []Project{
 			{
 				Name: "foo/project",
@@ -168,6 +171,7 @@ projects:
 		OnInitFetchRefsFromPipelines:           false,
 		OnInitFetchRefsFromPipelinesDepthLimit: defaultOnInitFetchRefsFromPipelinesDepthLimit,
 		DefaultRefsRegexp:                      "",
+		MaximumProjectsPollingWorkers:          runtime.GOMAXPROCS(0),
 		Projects: []Project{
 			{
 				Name: "foo/bar",
@@ -212,4 +216,41 @@ projects:
 	cfg.MergeWithContext(ctx)
 
 	assert.Equal(t, expectedCtxToken, cfg.Gitlab.Token)
+}
+
+func TestParsePrometheusConfig(t *testing.T) {
+	f, err := ioutil.TempFile("/tmp", "test-")
+	assert.Nil(t, err)
+	defer os.Remove(f.Name())
+
+	// Valid minimal configuration
+	f.WriteString(`
+prometheus_openmetrics_encoding: true
+projects:
+  - name: foo/project
+  - name: bar/project
+    refs: "^master|dev$"
+`)
+
+	config := &Config{}
+	assert.NoError(t, config.Parse(f.Name()))
+	assert.True(t, config.PrometheusOpenmetricsEncoding)
+}
+
+func TestParseConfigWithoutProjectWorkersUsesGOMAXPROCS(t *testing.T) {
+	f, err := ioutil.TempFile("/tmp", "test-")
+	assert.Nil(t, err)
+	defer os.Remove(f.Name())
+
+	// Valid minimal configuration
+	f.WriteString(`
+projects:
+  - name: foo/project
+  - name: bar/project
+    refs: "^master|dev$"
+`)
+	config := &Config{}
+	assert.NoError(t, config.Parse(f.Name()))
+	assert.Equal(t, runtime.GOMAXPROCS(0), config.MaximumProjectsPollingWorkers)
+
 }